+++ src/ViewRenderer.php
@@ -160,10 +160,9 @@ final class ViewRenderer implements ViewContextInterface
      *
      * @psalm-param array<string, mixed> $parameters
      *
-     * @throws RuntimeException If the view cannot be resolved.
      * @throws Throwable If an error occurred during rendering.
      * @throws ViewNotFoundException If the view file does not exist.
-     *
+     * @throws RuntimeException If the view cannot be resolved.
      * @return string The rendering result.
      */
     public function renderAsString(string $view, array $parameters = []): string
@@ -187,10 +186,9 @@ final class ViewRenderer implements ViewContextInterface
      *
      * @psalm-param array<string, mixed> $parameters
      *
-     * @throws RuntimeException If the view cannot be resolved.
      * @throws Throwable If an error occurred during rendering.
      * @throws ViewNotFoundException If the view file does not exist.
-     *
+     * @throws RuntimeException If the view cannot be resolved.
      * @return string The rendering result.
      */
     public function renderPartialAsString(string $view, array $parameters = []): string
@@ -306,7 +304,6 @@ final class ViewRenderer implements ViewContextInterface
      * @throws RuntimeException If the view cannot be resolved.
      * @throws Throwable If an error occurred during rendering.
      * @throws ViewNotFoundException If the view file does not exist.
-     *
      * @return string The rendering result.
      */
     private function renderProxy(
@@ -361,10 +358,10 @@ final class ViewRenderer implements ViewContextInterface
     private function getCommonParameters(): array
     {
         $parameters = [];
-        foreach ($this->getInjections(CommonParametersInjectionInterface::class) as $injection) {
-            $parameters = array_merge($parameters, $injection->getCommonParameters());
+        foreach ($this->getInjections($this->layout, CommonParametersInjectionInterface::class) as $injection) {
+            $parameters[] = $injection->getCommonParameters();
         }
-        return $parameters;
+        return array_merge(...$parameters);
     }
 
     /**
@@ -377,10 +374,10 @@ final class ViewRenderer implements ViewContextInterface
     private function getLayoutParameters(): array
     {
         $parameters = [];
-        foreach ($this->getInjections(LayoutParametersInjectionInterface::class) as $injection) {
-            $parameters = array_merge($parameters, $injection->getLayoutParameters());
+        foreach ($this->getInjections($this->layout, LayoutParametersInjectionInterface::class) as $injection) {
+            $parameters[] = $injection->getLayoutParameters();
         }
-        return $parameters;
+        return array_merge(...$parameters);
     }
 
     /**
@@ -391,10 +388,10 @@ final class ViewRenderer implements ViewContextInterface
     private function getMetaTags(): array
     {
         $tags = [];
-        foreach ($this->getInjections(MetaTagsInjectionInterface::class) as $injection) {
-            $tags = array_merge($tags, $injection->getMetaTags());
+        foreach ($this->getInjections($this->layout, MetaTagsInjectionInterface::class) as $injection) {
+            $tags[] = $injection->getMetaTags();
         }
-        return $tags;
+        return array_merge(...$tags);
     }
 
     /**
@@ -405,10 +402,10 @@ final class ViewRenderer implements ViewContextInterface
     private function getLinkTags(): array
     {
         $tags = [];
-        foreach ($this->getInjections(LinkTagsInjectionInterface::class) as $injection) {
-            $tags = array_merge($tags, $injection->getLinkTags());
+        foreach ($this->getInjections($this->layout, LinkTagsInjectionInterface::class) as $injection) {
+            $tags[] = $injection->getLinkTags();
         }
-        return $tags;
+        return array_merge(...$tags);
     }
 
     /**
@@ -416,12 +413,20 @@ final class ViewRenderer implements ViewContextInterface
      * @psalm-param class-string<T> $injectionInterface
      * @psalm-return list<T>
      */
-    private function getInjections(string $injectionInterface): array
+    private function getInjections(?string $layout, string $injectionInterface): array
     {
         $result = [];
         foreach ($this->getPreparedInjections() as $injection) {
             if ($injection instanceof $injectionInterface) {
                 $result[] = $injection;
+                continue;
+            }
+            if ($injection instanceof LayoutSpecificInjections && $injection->getLayout() === $layout) {
+                foreach ($injection->getInjections() as $layoutInjection) {
+                    if ($layoutInjection instanceof $injectionInterface) {
+                        $result[] = $layoutInjection;
+                    }
+                }
             }
         }
         return $result;